Added effects folder and effect_menu.py
