+++ effects/anime_effect.py
