+++ effects/cartoon_effects.py
