+++ effects/comic_cartoon_effect.py
