+++ effects/mosaic_creator.py
