+++ effects/oil_painting.py
