+++ effects/pencil_sketch.py
