+++ effects/stipple.py
