+++ requirements.txt
@@ -5,3 +5,4 @@ pillow
 scipy
 scikit-learn
 streamlit
+tqdm
